backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pysqlcipher3>=1.2.0
pytest>=7.4.0
httpx>=0.26.0
mnemonic>=0.21
argon2-cffi>=23.1.0
cryptography>=41.0.0
